requirements.txt
datasets
pandas
chromadb
streamlit
requests 
python-dotenv
redis
sentence-tranformers